% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{attach_window}
\alias{attach_window}
\title{Attach to an existing tmux window}
\usage{
attach_window(x, lookup_id = TRUE)
}
\arguments{
\item{x}{An integer or string indicating the name of an existing window.}

\item{lookup_id}{A logical. If \code{FALSE}, \code{x} is assumed to be a window id.
Default: \code{TRUE}.}
}
\value{
A tmuxr_window.
}
\description{
Attach to an existing tmux window
}
