% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{capture_pane}
\alias{capture_pane}
\title{Capture the contents of a pane.}
\usage{
capture_pane(
  target,
  start = NULL,
  end = NULL,
  escape = FALSE,
  escape_control = FALSE,
  join = FALSE,
  cat = FALSE
)
}
\arguments{
\item{target}{A session, window, or pane.}

\item{start}{Numeric. First line to capture.}

\item{end}{Numeric. Last line to capture.}

\item{escape}{Logical. Include escape sequences for text and background attributes.}

\item{escape_control}{Logical. Also escape control characters as octal \\xxx.}

\item{join}{Logical. Join wrapped lines and preserve trailing spaces at each line.}

\item{cat}{Logical. Concatenate lines into one string.}
}
\description{
Capture the contents of a pane.
}
