% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{clock_mode}
\alias{clock_mode}
\title{Display a large clock}
\usage{
clock_mode(target = NULL)
}
\arguments{
\item{target}{A \code{tmuxr_session}, \code{tmuxr_window}, or \code{tmuxr_pane}.}
}
\description{
Display a large clock
}
