% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/props.R
\name{display_message}
\alias{display_message}
\title{Display a message}
\usage{
display_message(target = NULL, message = NULL, verbose = FALSE, stdout = TRUE)
}
\arguments{
\item{target}{A session, window, or pane.}

\item{message}{The message to display. Refer to the FORMATS section of the tmux man page for the format.}

\item{verbose}{Logical. Print verbose logging as the format is parsed?}

\item{stdout}{Logical. Print message to stardard output?}
}
\description{
Display a message
}
