% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{get_prompt}
\alias{get_prompt}
\title{Get the prompt pattern of a session, window or pane.}
\usage{
get_prompt(target)
}
\arguments{
\item{target}{A session, window, or pane.}
}
\value{
String containing a regular expression that matches all relevant
prompts.
}
\description{
Get the prompt pattern of a session, window or pane.
}
