% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{has_prompt}
\alias{has_prompt}
\title{Is there a prompt pattern associated with a session, window, or pane?}
\usage{
has_prompt(target)
}
\arguments{
\item{target}{A session, window, or pane.}
}
\description{
Is there a prompt pattern associated with a session, window, or pane?
}
