% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{id}
\alias{id}
\title{Identifier of a tmux object}
\usage{
id(target)
}
\arguments{
\item{target}{A tmuxr_session, tmuxr_window, or tmuxr_pane.}
}
\value{
A string.
}
\description{
Identifier of a tmux object
}
