% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{is_active}
\alias{is_active}
\title{Is tmux object active?}
\usage{
is_active(target)
}
\arguments{
\item{target}{A tmuxr_window or tmuxr_pane.}
}
\value{
A logical.
}
\description{
Is tmux object active?
}
