% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{kill_pane}
\alias{kill_pane}
\title{Kill a tmux pane}
\usage{
kill_pane(target = NULL, inverse = FALSE)
}
\arguments{
\item{target}{A tmuxr_pane. If \code{NULL}, the currently active pane is used.
Default: \code{NULL}.}

\item{inverse}{A logical. If \code{TRUE}, kills all but the \code{target} pane.
Default: \code{FALSE}.}
}
\description{
Destroy the given pane. If no panes remain in the containing window, it is
also destroyed.
}
