% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{kill_session}
\alias{kill_session}
\title{Kill a tmux session}
\usage{
kill_session(target, inverse = FALSE)
}
\arguments{
\item{target}{A tmuxr_session.}

\item{inverse}{A logical. If \code{TRUE}, kills all but the \code{target} session.
Default: \code{FALSE}.}
}
\description{
Kill a tmux session
}
