% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{kill_window}
\alias{kill_window}
\title{Kill a tmux window}
\usage{
kill_window(target, inverse = FALSE)
}
\arguments{
\item{target}{A tmuxr_window.}

\item{inverse}{A logical. If \code{TRUE}, kills all but the \code{target} window.
Default: \code{FALSE}.}
}
\description{
Kill a tmux window
}
