% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout}
\alias{layout}
\alias{layout<-}
\alias{set_layout}
\alias{layout_even_horizontal}
\alias{layout_even_vertical}
\alias{layout_main_horizontal}
\alias{layout_main_vertical}
\alias{layout_tiled}
\alias{layout_next}
\alias{layout_previous}
\alias{layout_even}
\alias{layout_undo}
\alias{layout_rotate}
\title{Change layout of a tmux window}
\usage{
layout(target = NULL)

layout(target) <- value

set_layout(target = NULL, value)

layout_even_horizontal(target = NULL)

layout_even_vertical(target = NULL)

layout_main_horizontal(target = NULL, height = NULL)

layout_main_vertical(target = NULL, width = NULL)

layout_tiled(target = NULL)

layout_next(target = NULL)

layout_previous(target = NULL)

layout_even(target = NULL)

layout_undo(target = NULL)

layout_rotate(target = NULL, reverse = FALSE)
}
\arguments{
\item{target}{A tmuxr_window. If \code{NULL}, use currently active window.}

\item{value}{A string. Layout definition or name of preset.}

\item{width, height}{An integer. Width or height of the main pane.}

\item{reverse}{A logical. If \code{TRUE} the direction in which rotate the window
is reversed. Default: \code{FALSE}.}
}
\value{
A string. Layout definition.
}
\description{
Functions to get and set the layout of a tmux window.
}
