% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{list_sessions}
\alias{list_sessions}
\title{List sessions.}
\usage{
list_sessions()
}
\value{
A list of \code{tmuxr_session}s.
}
\description{
List sessions.
}
\examples{
\dontshow{try(tmuxr::kill_server(), silent = TRUE)}
a <- new_session('first')
b <- new_session('second')
list_sessions()
list_sessions() \%>\%
  purrr::walk(kill_session)
\dontshow{try(tmuxr::kill_server(), silent = TRUE)}
}
