% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{new_session}
\alias{new_session}
\title{Create a new session}
\usage{
new_session(name = NULL, prompt = prompts$bash, window_name = NULL,
  start_directory = NULL, width = NULL, height = NULL, detached = TRUE,
  shell_command = NULL)
}
\description{
Create a new session
}
