% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/props.R
\name{object_name}
\alias{object_name}
\alias{name}
\alias{name<-}
\title{Name of a tmux object}
\usage{
name(target)

name(target) <- value
}
\arguments{
\item{target}{A \code{tmuxr_session}, \code{tmuxr_window}, or \code{tmuxr_pane}.}

\item{value}{String.}
}
\description{
Name of a tmux object
}
\seealso{
\code{\link[=rename_session]{rename_session()}}, \code{\link[=rename_window]{rename_window()}}
}
