% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{pane_from_name}
\alias{pane_from_name}
\title{Create a \code{tmuxr_pane} from an existing tmux pane.}
\usage{
pane_from_name(name)
}
\arguments{
\item{name}{Numeric or string indicating the name of the existing pane.}
}
\value{
A \code{tmuxr_pane}.
}
\description{
Create a \code{tmuxr_pane} from an existing tmux pane.
}
