% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{pipe_pane}
\alias{pipe_pane}
\title{Pipe contents of a tmux pane to a shell command}
\usage{
pipe_pane(
  target = NULL,
  shell_command = NULL,
  stdout = TRUE,
  stdin = FALSE,
  open = FALSE
)
}
\arguments{
\item{target}{A \code{tmuxr_session}, \code{tmuxr_window}, or \code{tmuxr_pane}.}

\item{shell_command}{String. If \code{NULL}, the current pipe (if any) is closed.}

\item{stdout}{Logical. Connect standard input of pane to \code{shell_command}?}

\item{stdin}{Logical. Connect standard output of pane to \code{shell_command}?}

\item{open}{Logical. Only open a new pipe if no previous pipe exists.}
}
\description{
Pipe contents of a tmux pane to a shell command
}
