% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\docType{data}
\name{prompts}
\alias{prompts}
\title{A list of commonly used prompt patterns.}
\format{An object of class \code{list} of length 6.}
\usage{
prompts
}
\description{
A list of commonly used prompt patterns.
}
\keyword{datasets}
