% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop}
\alias{prop}
\title{Properties of a tmux object}
\usage{
prop(target = NULL, property)
}
\arguments{
\item{target}{A tmuxr_session, tmuxr_window, or tmuxr_session. Default:
\code{NULL}.}

\item{property}{A string.}
}
\value{
A string.
}
\description{
Get the properties of a tmux object, including sessions, windows,
and panes.
}
\examples{
s <- new_session()
prop(s, "session_created")
kill_session(s)

}
