% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{rename_session}
\alias{rename_session}
\title{Rename a session.}
\usage{
rename_session(target, new_name)
}
\arguments{
\item{new_name}{String indicating the new name of the session}

\item{session}{A \code{tmuxr_session}.}
}
\value{
A \code{tmuxr_session}.
}
\description{
Rename a session.
}
