% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{rename_window}
\alias{rename_window}
\title{Rename a tmux window}
\usage{
rename_window(target, value)
}
\arguments{
\item{target}{A \code{tmuxr_window}.}

\item{value}{String indicating the new name of the window.}
}
\value{
A \code{tmuxr_window}.
}
\description{
Rename a tmux window
}
