% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{resize_pane}
\alias{resize_pane}
\title{Resize pane.}
\usage{
resize_pane(target, width = NULL, height = NULL)
}
\arguments{
\item{target}{A pane.}

\item{width}{Numeric.}

\item{height}{Numeric.}
}
\description{
Resize pane.
}
