% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{resize_window}
\alias{resize_window}
\title{Resize window.}
\usage{
resize_window(target, width = NULL, height = NULL)
}
\arguments{
\item{target}{A session.}

\item{width}{Numeric.}

\item{height}{Numeric.}
}
\description{
Resize window.
}
