% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_pane}
\alias{select_pane}
\alias{select_pane_active}
\alias{select_pane_last}
\alias{select_pane_next}
\alias{select_pane_previous}
\alias{select_pane_top}
\alias{select_pane_bottom}
\alias{select_pane_left}
\alias{select_pane_right}
\alias{select_pane_top_left}
\alias{select_pane_top_right}
\alias{select_pane_bottom_left}
\alias{select_pane_bottom_right}
\alias{select_pane_up_of}
\alias{select_pane_down_of}
\alias{select_pane_left_of}
\alias{select_pane_right_of}
\title{Select a tmux pane}
\usage{
select_pane(target = NULL, token = NULL)

select_pane_active(target = NULL)

select_pane_last(target = NULL)

select_pane_next(target = NULL)

select_pane_previous(target = NULL)

select_pane_top(target = NULL)

select_pane_bottom(target = NULL)

select_pane_left(target = NULL)

select_pane_right(target = NULL)

select_pane_top_left(target = NULL)

select_pane_top_right(target = NULL)

select_pane_bottom_left(target = NULL)

select_pane_bottom_right(target = NULL)

select_pane_up_of(target = NULL)

select_pane_down_of(target = NULL)

select_pane_left_of(target = NULL)

select_pane_right_of(target = NULL)
}
\arguments{
\item{target}{A tmuxr_session, tmuxr_window, or tmuxr_pane. If \code{NULL}, the
currently active pane is used. Default: \code{NULL}.}

\item{token}{A string. Default \code{NULL}.}
}
\value{
A tmuxr_pane.
}
\description{
Note that selecting a pane from a non-active window does not select that
window.
}
