% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_window}
\alias{select_window}
\alias{select_window_active}
\alias{select_window_last}
\alias{select_window_next}
\alias{select_window_previous}
\alias{select_window_start}
\alias{select_window_end}
\title{Select a tmux window}
\usage{
select_window(target = NULL, token = NULL)

select_window_active(target = NULL)

select_window_last(target = NULL)

select_window_next(target = NULL)

select_window_previous(target = NULL)

select_window_start(target = NULL)

select_window_end(target = NULL)
}
\arguments{
\item{target}{A tmuxr_session or tmuxr_window. If \code{NULL}, the
last session is used. Default: \code{NULL}.}

\item{token}{A string. Default \code{NULL}.}
}
\value{
A tmuxr_window.
}
\description{
Select a tmux window
}
