% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_backspace}
\alias{send_backspace}
\title{Send the Backspace key to a session, window or pane.}
\usage{
send_backspace(target)
}
\arguments{
\item{target}{A session, window, or pane.}
}
\description{
Send the Backspace key to a session, window or pane.
}
