% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_control_c}
\alias{send_control_c}
\title{Send the Control C combination to a session, window or pane.}
\usage{
send_control_c(target)
}
\arguments{
\item{target}{A session, window, or pane.}
}
\description{
Send the Control C combination to a session, window or pane.
}
