% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_enter}
\alias{send_enter}
\title{Send the Enter key to a session, window or pane.}
\usage{
send_enter(target)
}
\arguments{
\item{target}{A session, window, or pane.}
}
\description{
Send the Enter key to a session, window or pane.
}
