% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_keys}
\alias{send_keys}
\title{Send keys to a session, window, or pane.}
\usage{
send_keys(target, keys, literal = FALSE)
}
\arguments{
\item{target}{A \code{tmuxr_session}, \code{tmuxr_window}, or \code{tmuxr_pane}.}

\item{keys}{String to send.}

\item{literal}{Should the keys be interpreted literally?}
}
\description{
Send keys to a session, window, or pane.
}
