% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_lines}
\alias{send_lines}
\title{Send multiple lines to a session, window, or pane.}
\usage{
send_lines(target, lines, wait = TRUE)
}
\arguments{
\item{target}{A session, window, or pane.}

\item{lines}{A character vector.}

\item{wait}{Should there be waited for the prompt after sending each line?}
}
\description{
Send multiple lines to a session, window, or pane.
}
