% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_prefix}
\alias{send_prefix}
\title{Send prefix to a tmux pane}
\usage{
send_prefix(target = NULL, secondary = FALSE)
}
\arguments{
\item{target}{A tmuxr_session, tmuxr_window, or tmuxr_pane. If \code{NULL},
the currently active pane is used. Default: \code{NULL}.}

\item{secondary}{A logical. If \code{TRUE}, send secondary prefix. Default:
\code{FALSE}.}
}
\description{
Send prefix to a tmux pane
}
