% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session_from_name}
\alias{session_from_name}
\title{Create a \code{tmuxr_session} from an existing tmux session.}
\usage{
session_from_name(name, prompt = prompts$bash)
}
\arguments{
\item{name}{Numeric or string indicating the name of the existing session.}

\item{prompt}{String containing a regular expression.}
}
\value{
A \code{tmuxr_session}.
}
\description{
Create a \code{tmuxr_session} from an existing tmux session.
}
\examples{
\dontshow{kill_server()}
new_session()
s <- session_from_name(0)
}
