% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_option}
\alias{set_option}
\title{Set option}
\usage{
set_option(
  target = NULL,
  option,
  value = NULL,
  type = c("session", "window", "pane", "server"),
  append = FALSE,
  expand = FALSE,
  global = FALSE,
  unset = FALSE,
  override = TRUE
)
}
\arguments{
\item{target}{A tmuxr_session, tmuxr_window, or tmuxr_pane. Default: \code{NULL}.}

\item{option}{A string.}

\item{value}{A string. Default: \code{NULL}.}

\item{type}{A string. Type of option. Default: \code{session}.}

\item{append}{A logical. Default: \code{FALSE}.}

\item{expand}{A logical. Default: \code{FALSE}.}

\item{global}{A logical. Default: \code{FALSE}.}

\item{unset}{A logical. Default: \code{FALSE}.}

\item{override}{A logical. Default: \code{TRUE}.}
}
\description{
Set option
}
