% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{set_prompt}
\alias{set_prompt}
\title{Set the prompt pattern of a session, window or pane.}
\usage{
set_prompt(target, prompt)
}
\arguments{
\item{target}{A session, window, or pane.}

\item{prompt}{String containing a regular expression that matches all
relevant prompts.}
}
\description{
Set the prompt pattern of a session, window or pane.
}
