% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{style}
\alias{style}
\alias{style<-}
\alias{set_style}
\title{Style of a tmux window or pane}
\usage{
style(target)

style(target) <- value

set_style(target, value)
}
\arguments{
\item{target}{A tmuxr_window or tmuxr_pane.}

\item{value}{A named list. Style definition.}
}
\description{
Functions to get and set the layout of a tmux window or pane.
}
