% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{style_convert}
\alias{style_convert}
\alias{strfstyle}
\alias{strpstyle}
\title{Format and parse style strings}
\usage{
strfstyle(...)

strpstyle(x)
}
\arguments{
\item{...}{Named strings and logicals or one named list.
Colors and attributes.}

\item{x}{A string.}
}
\value{
A string or named list.
}
\description{
Functions to convert a style string to a named list and back.
}
\examples{
\dontrun{
strfstyle(fg = "red", bg = "#00ff00", blink = TRUE, align = FALSE)
strpstyle("fg=red,bg=#00ff00,blink,noalign")
}
}
