% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{swap_pane}
\alias{swap_pane}
\alias{swap_window}
\title{Swap two tmux windows or panes}
\usage{
swap_pane(from = NULL, to = NULL, reverse = FALSE, select = TRUE)

swap_window(from = NULL, to = NULL, select = TRUE)
}
\arguments{
\item{from}{A tmuxr_window or tmuxr_pane.}

\item{to}{A tmuxr_window or tmuxr_pane.}

\item{reverse}{A logical.}

\item{select}{A logical. If \code{FALSE}, do not select the new window or pane.
Default: \code{TRUE}.}
}
\description{
Swap two tmux windows or panes
}
