% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmux.R
\name{tmux_command}
\alias{tmux_command}
\title{Execute a tmux command}
\usage{
tmux_command(command, ...)
}
\arguments{
\item{command}{String.}

\item{...}{Character vector of command flags.}
}
\description{
Execute a tmux command
}
