% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmuxr-package.R
\docType{package}
\name{tmuxr-package}
\alias{tmuxr}
\alias{tmuxr-package}
\title{tmuxr: Manage 'tmux' Sessions, Windows, and Panes}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Create, control, and capture 'tmux' sessions, windows, and panes
    using a pipeable API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://datascienceworkshops.github.io/tmuxr}
  \item \url{https://github.com/datascienceworkshops/tmuxr}
  \item Report bugs at \url{https://github.com/datascienceworkshops/tmuxr/issues}
}

}
\author{
\strong{Maintainer}: Jeroen Janssens \email{jeroen@datascienceworkshops.com} (\href{https://orcid.org/0000-0002-5878-8924}{ORCID})

Other contributors:
\itemize{
  \item Data Science Workshops [copyright holder, funder]
}

}
\keyword{internal}
