% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{wait}
\alias{wait}
\title{Wait.}
\usage{
wait(target, time = 0.05)
}
\arguments{
\item{target}{A session, window, or pane.}

\item{time}{Numerical. Time to wait in seconds between tries.}
}
\description{
Wait.
}
