% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_from_name}
\alias{window_from_name}
\title{Create a \code{tmuxr_window} from an existing tmux window.}
\usage{
window_from_name(name)
}
\arguments{
\item{name}{Numeric or string indicating the name of the existing window.}
}
\value{
A \code{tmuxr_window}.
}
\description{
Create a \code{tmuxr_window} from an existing tmux window.
}
