% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{attach_pane}
\alias{attach_pane}
\title{Attach to an existing tmux pane.}
\usage{
attach_pane(name)
}
\arguments{
\item{name}{Numeric or string indicating the name of the existing pane.}
}
\value{
A \code{tmuxr_pane}.
}
\description{
Attach to an existing tmux pane.
}
