% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{attach_pane}
\alias{attach_pane}
\title{Attach to an existing tmux pane}
\usage{
attach_pane(x, lookup_id = TRUE)
}
\arguments{
\item{x}{An integer or string indicating the name of an existing pane.}

\item{lookup_id}{A logical. If \code{FALSE}, \code{x} is assumed to be a pane id.
Default: \code{TRUE}.}
}
\value{
A tmuxr_pane.
}
\description{
Attach to an existing tmux pane
}
