% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{attach_session}
\alias{attach_session}
\title{Attach to an existing tmux session}
\usage{
attach_session(x, lookup_id = TRUE)
}
\arguments{
\item{x}{An integer or string indicating the name of an existing window.}

\item{lookup_id}{A logical. If \code{FALSE}, \code{x} is assumed to be a session id.
Default: \code{TRUE}.}
}
\value{
A tmuxr_session.
}
\description{
Attach to an existing tmux session
}
