% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{attach_session}
\alias{attach_session}
\title{Attach to an existing tmux session.}
\usage{
attach_session(name, prompt = NULL)
}
\arguments{
\item{name}{Numeric or string indicating the name of the existing session.}

\item{prompt}{String containing a regular expression.}
}
\value{
A \code{tmuxr_session}.
}
\description{
Attach to an existing tmux session.
}
