% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{attach_window}
\alias{attach_window}
\title{Attach to an existing tmux window.}
\usage{
attach_window(name)
}
\arguments{
\item{name}{Numeric or string indicating the name of the existing window.}
}
\value{
A \code{tmuxr_window}.
}
\description{
Attach to an existing tmux window.
}
