% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{capture_pane}
\alias{capture_pane}
\title{Capture contents of a pane.}
\usage{
capture_pane(target, start = NULL, end = NULL, trim = FALSE,
  trim_top = trim, trim_bottom = trim, strip_lonely_prompt = TRUE,
  as_message = FALSE, in_viewer = FALSE)
}
\description{
Capture contents of a pane.
}
