% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{clock_mode}
\alias{clock_mode}
\title{Display a large clock}
\usage{
clock_mode(target = NULL)
}
\arguments{
\item{target}{A tmuxr_session, tmuxr_window, or tmuxr_pane. If \code{NULL}, the
currently active pane is used. Default: \code{NULL}.}
}
\description{
Display a large clock
}
