% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{kill_server}
\alias{kill_server}
\title{Kill the tmux server}
\usage{
kill_server(silent = TRUE)
}
\arguments{
\item{silent}{A logical. Should a possible error be suppressed?
Default: \code{TRUE}.}
}
\description{
Kill the tmux server
}
