% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{list_panes}
\alias{list_panes}
\title{List tmux panes}
\usage{
list_panes(target = NULL)
}
\arguments{
\item{target}{A tmuxr_session or tmuxr_window. The parent session or
window. If \code{NULL} all panes of all sessions are listed. Default: \code{NULL}.}
}
\value{
A list of tmuxr_panes.
}
\description{
List tmux panes
}
