% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{list_panes}
\alias{list_panes}
\title{List panes.}
\usage{
list_panes(target = NULL)
}
\arguments{
\item{target}{Parent session or window. If \code{NULL} (default), all windows are
listed.}
}
\value{
A list of \code{tmuxr_pane}s.
}
\description{
List panes.
}
