% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{list_windows}
\alias{list_windows}
\title{List tmux windows}
\usage{
list_windows(target = NULL)
}
\arguments{
\item{target}{Name of parent tmuxr_session. If \code{NULL} (default), all
windows of all sessions are listed.}
}
\value{
A list of tmuxr_windows.
}
\description{
List tmux windows
}
