% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{list_windows}
\alias{list_windows}
\title{List windows.}
\usage{
list_windows(target = NULL)
}
\arguments{
\item{target}{Name of parent session. If \code{NULL} (default), all windows are
listed.}
}
\value{
A list of \code{tmuxr_window}s.
}
\description{
List windows.
}
