% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{object_name}
\alias{object_name}
\alias{name}
\alias{name<-}
\alias{set_name}
\title{Name of a tmux object}
\usage{
name(target)

name(target) <- value

set_name(target, value)
}
\arguments{
\item{target}{A tmuxr_session, tmuxr_window, or tmuxr_pane.}

\item{value}{A string. New name.}
}
\value{
A string.
}
\description{
Functions to get and set the name of a tmux session, window, and pane.
}
\seealso{
\code{\link[=rename_session]{rename_session()}}, \code{\link[=rename_window]{rename_window()}}
}
