% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/props.R
\name{object_size}
\alias{object_size}
\alias{width}
\alias{width<-}
\alias{height}
\alias{height<-}
\title{Width and height of a tmux object}
\usage{
width(target)

width(target) <- value

height(target)

height(target) <- value
}
\arguments{
\item{target}{A \code{tmuxr_session}, \code{tmuxr_window}, or \code{tmuxr_pane}.}

\item{value}{Numeric.}
}
\description{
Width and height of a tmux object
}
\seealso{
\code{\link[=resize_window]{resize_window()}}, \code{\link[=resize_pane]{resize_pane()}}
}
