% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{pipe_pane}
\alias{pipe_pane}
\title{Pipe contents of a pane to a shell command.}
\usage{
pipe_pane(target = NULL, shell_command = NULL, open = FALSE)
}
\description{
Pipe contents of a pane to a shell command.
}
