% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{pipe_pane}
\alias{pipe_pane}
\title{Pipe contents of a tmux pane to a shell command}
\usage{
pipe_pane(
  target = NULL,
  shell_command = NULL,
  stdout = TRUE,
  stdin = FALSE,
  open = FALSE
)
}
\arguments{
\item{target}{A tmuxr_session, tmuxr_window, or tmuxr_pane.}

\item{shell_command}{String. If \code{NULL}, the current pipe (if any) is closed.
Default: \code{NULL}.}

\item{stdout}{Logical. Connect standard output of pane to \code{shell_command}?
Default: \code{TRUE}.}

\item{stdin}{Logical. Connect standard input of pane to \code{shell_command}?
Default: \code{FALSE}.}

\item{open}{Logical. Only open a new pipe if no previous pipe exists.
Default: \code{FALSE}.}
}
\description{
Pipe contents of a tmux pane to a shell command
}
