% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{pipe_pane}
\alias{pipe_pane}
\title{Pipe contents of a pane to a shell command.}
\usage{
pipe_pane(target = NULL, shell_command = NULL, open = FALSE)
}
\arguments{
\item{target}{A session, window, or pane.}

\item{shell_command}{String. System command to be invoked when creating the
session.}

\item{open}{Logical. Only open a new pipe if no previous pipe exists.}
}
\description{
Pipe contents of a pane to a shell command.
}
