% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{rename_session}
\alias{rename_session}
\title{Rename a tmux session}
\usage{
rename_session(target, value)
}
\arguments{
\item{target}{A \code{tmuxr_session}.}

\item{value}{String indicating the new name of the session.}
}
\value{
A \code{tmuxr_session}.
}
\description{
Rename a tmux session
}
