% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{rename_window}
\alias{rename_window}
\title{Rename a window}
\usage{
rename_window(target, new_name)
}
\arguments{
\item{target}{A \code{tmuxr_window}.}

\item{new_name}{String indicating the new name of the session}
}
\value{
A \code{tmuxr_window}.
}
\description{
Rename a window
}
