% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{resize_pane}
\alias{resize_pane}
\title{Resize a tmux pane}
\usage{
resize_pane(target, width = NULL, height = NULL)
}
\arguments{
\item{target}{A \code{tmuxr_pane}.}

\item{width}{Numeric.}

\item{height}{Numeric.}
}
\description{
Resize a tmux pane
}
