% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{resize_pane}
\alias{resize_pane}
\title{Resize a tmux pane}
\usage{
resize_pane(target, width = NULL, height = NULL)
}
\arguments{
\item{target}{A tmuxr_pane.}

\item{width}{An integer. Default: \code{NULL}.}

\item{height}{An integer. Default: \code{NULL}.}
}
\description{
Resize a tmux pane
}
