% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{resize_window}
\alias{resize_window}
\title{Resize a tmux window}
\usage{
resize_window(target, width = NULL, height = NULL)
}
\arguments{
\item{target}{A tmuxr_session or tmuxr_window.}

\item{width}{An integer. Default: \code{NULL}.}

\item{height}{An integer. Default: \code{NULL}.}
}
\description{
Resize a tmux window
}
