% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmux_command}
\alias{tmux_command}
\title{Execute a tmux command}
\usage{
tmux_command(command, ..., .silent = FALSE)
}
\arguments{
\item{command}{A string.}

\item{...}{Strings. Command flags.}

\item{.silent}{A Logical. Default: \code{FALSE}.}
}
\value{
A vector of strings.
}
\description{
Execute a tmux command
}
\examples{
\dontrun{
s <- new_session()
split_window(s)
tmux_command("list-panes", "-t" , "0")
kill_session(s)
}

}
