% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{tmux_version}
\alias{tmux_version}
\title{Get tmux version}
\usage{
tmux_version(as_numeric = TRUE)
}
\arguments{
\item{as_numeric}{A logical. If \code{TRUE}, the version number is returned as a
numeric. Default: \code{TRUE}.}
}
\value{
A numeric or string.
}
\description{
Get the version of tmux.
}
\examples{
\dontrun{
tmux_version()
tmux_version() >= 2.8
tmux_version(as_numeric = FALSE)
}

}
