% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{wait_for_prompt}
\alias{wait_for_prompt}
\title{Wait for prompt to appear in session, window or pane.}
\usage{
wait_for_prompt(target, time = 0.05)
}
\arguments{
\item{target}{A session, window, or pane.}

\item{time}{Numerical. Time to wait in seconds between tries.}
}
\description{
Wait for prompt to appear in session, window or pane.
}
